requirements.txt
asgiref==3.3.0
Django==3.0
django-braces==1.14.0
django-cors-headers==3.5.0
django-model-utils==4.0.0
django-star-ratings==0.9.1
djangorestframework==3.12.1
Pillow==8.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
